massive refactoring effort part1 - getting closer (see refactor.documentation for details)
