+++ src/js/util/entitydata.js
@@ -1,0 +1,178 @@
+//////// Module Definition ////////////
+define([
+	'util/util', // pulls in angular
+	'util/wikidataapi',
+	'util/i18n'
+], function() {
+///////////////////////////////////////
+
+angular.module('utilities').factory('entitydata', ['wikidataapi', 'util', 'i18n', function(wikidataapi, util, i18n) {
+
+	var getStatementValue = function(statementJson, defaultValue) {
+		try {
+			var ret = statementJson.mainsnak.datavalue.value;
+			if (ret) return ret;
+		} catch (err) {
+			// fall through
+		}
+		return defaultValue;
+	}
+
+	var addEntityIdsFromSnak = function(snak, missingIds) {
+		if (snak.snaktype == 'value') {
+			switch (snak.datavalue.type) {
+				case 'wikibase-entityid':
+					if (snak.datavalue.value["entity-type"] == "item") {
+						missingIds["Q" + snak.datavalue.value["numeric-id"]] = true;
+					}
+					break;
+				case 'quantity':
+					var unit = util.getIdFromUri(snak.datavalue.value.unit);
+					if (unit !== null) {
+						missingIds[unit] = true;
+					}
+					break;
+				case 'globecoordinate':
+					var globe = util.getIdFromUri(snak.datavalue.value.globe);
+					if (globe !== null) {
+						missingIds[globe] = true;
+					}
+					break;
+				case 'time':
+				case 'string':
+				case 'monolingualtext':
+				default:
+					break; // no ids
+			}
+		}
+	}
+
+	var getEntityIds = function(statements) {
+		var result = {};
+		angular.forEach(statements, function(statementGroup) {
+			angular.forEach(statementGroup, function (statement) {
+				addEntityIdsFromSnak(statement.mainsnak, result);
+				if ('qualifiers' in statement) {
+					angular.forEach(statement.qualifiers, function (snakList) {
+						angular.forEach(snakList, function(snak) {
+							addEntityIdsFromSnak(snak, result);
+						});
+					});
+				}
+			});
+		});
+		return Object.keys(result);
+	}
+
+	var addPropertyIdsFromSnak = function(snak, missingIds) {
+		if ( snak.snaktype == 'value'
+			&& snak.datavalue.type == 'wikibase-entityid'
+			&& snak.datavalue.value["entity-type"] == "property" ) {
+				missingIds["P" + snak.datavalue.value["numeric-id"]] = true;
+		}
+		missingIds[snak.property] = true;
+	}
+
+	var getPropertyIds = function(statements) {
+		var result = {};
+		angular.forEach(statements, function(statementGroup) {
+			angular.forEach(statementGroup, function (statement) {
+				addPropertyIdsFromSnak(statement.mainsnak, result);
+				if ('qualifiers' in statement) {
+					angular.forEach(statement.qualifiers, function (snakList) {
+						angular.forEach(snakList, function(snak) {
+							addPropertyIdsFromSnak(snak, result);
+						});
+					});
+				}
+			});
+		});
+		return Object.keys(result);
+	}
+
+	var getEntityData = function(id) {
+		var language = i18n.getLanguage();
+		return wikidataapi.getEntityData(id, language).then(function(response) {
+			var ret = {
+				language: language, // this is fixed for this result!
+				label: '',
+				labelorid: id,
+				description: '',
+				images: [],
+				aliases: [],
+				banner: null,
+				homepage: null,
+				statements: {},
+				missing: false,
+				termsPromise: null,
+				propLabelPromise: null,
+				waitForPropertyLabels: function() {
+					if (this.propLabelPromise == null) {
+						var propIdList = getPropertyIds(this.statements);
+						this.propLabelPromise = i18n.waitForPropertyLabels(propIdList, language);
+					}
+					return this.propLabelPromise;
+				},
+				waitForTerms: function() {
+					if (this.termsPromise == null) {
+						var termIdList = getEntityIds(this.statements);
+						this.termsPromise = i18n.waitForTerms(termIdList, language);
+					}
+					return this.termsPromise;
+				}
+			};
+
+			if ("error" in response || "missing" in response.entities[id]) {
+				ret.missing = true;
+				return ret;
+			}
+
+			var entityData = response.entities[id];
+
+			if ("labels" in entityData && ret.language in entityData.labels) {
+				ret.label = entityData.labels[ret.language].value;
+				ret.labelorid = entityData.labels[ret.language].value;
+			}
+			if ("descriptions" in entityData && ret.language in entityData.descriptions) {
+				ret.description = entityData.descriptions[ret.language].value;
+			}
+			if ("aliases" in entityData && ret.language in entityData.aliases) {
+				var aliasesData = entityData.aliases[ret.language];
+				for (var i in aliasesData){
+					ret.aliases.push(aliasesData[i].value);
+				}
+			}
+
+			if ("claims" in entityData) {
+				// image
+				if ("P18" in entityData.claims) {
+					for (var i in entityData.claims.P18) {
+						var imageFileName = getStatementValue(entityData.claims.P18[i],"");
+						ret.images.push(imageFileName.replace(" ","_"));
+					}
+				}
+				// Wikivoyage banner; only pick the first banner if multiple
+				if ("P948" in entityData.claims) {
+					var imageFileName = getStatementValue(entityData.claims.P948[0],null);
+					ret.banner = imageFileName.replace(" ","_");
+				}
+				
+				// homepage URL; only pick the first URL if multiple
+				if ("P856" in entityData.claims) {
+					ret.homepage = getStatementValue(entityData.claims.P856[0],null);
+				}
+
+				ret.statements = entityData.claims;
+			}
+
+			return ret;
+		});
+	};
+	
+	return {
+		getEntityData: getEntityData
+	};
+}]);
+
+return {}; // module
+});		  // definition end
